PIFO: Higher than available priorities are trunked
When the desired priority is lower than the priority available in a PIFO
rotation, the PIFO will drop the packet and force the scheduler to track
the bounds. This change assigns the packet to the highest available
priority instead of dropping the packet.

In the future, it is better to configure how many priorities should
always be available backward in each rotation. One way would be to
always add an offset to the next PIFO equal to the configured range.
This offset would duplicate the range from the currently active PIFO. It
would be unused until the next rotation when the PIFO is live.

Signed-off-by: Frey Alfredsson <freysteinn@freysteinn.com>